exo1/exoplanet_siamese/requirements.txt
# Core dependencies
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
scipy==1.11.1

# Deep Learning
torch==2.0.1
torchvision==0.15.2

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0

# Data processing
tqdm==4.65.0

# Configuration
pyyaml==6.0

# Jupyter support
# jupyter==1.0.0
# notebook==7.0.0
# ipykernel==6.24.0

# Additional utilities
joblib==1.3.1
